adding safeLog submodule
